# openapi.yaml
openapi: 3.0.3
info:
  title: svc-subman
  version: 1.0.0
  contact: {}
servers:
  - url: localhost:8080
paths:
  /subscribe:
    post:
      summary: create subscribe
      description: create subscribe
      operationId: createSubscribe
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                end_date:
                  type: string
                  example: 12-2024
                service_name:
                  type: string
                  example: Netflix Premium
                start_date:
                  type: string
                  example: 01-2024
                user_id:
                  type: string
                  example: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
            examples:
              create subscribe:
                value:
                  end_date: 12-2024
                  service_name: Netflix Premium
                  start_date: 01-2024
                  user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
      responses:
        '200':
          description: sucess
          headers:
            Content-Length:
              schema:
                type: string
                example: '66'
            Date:
              schema:
                type: string
                example: Fri, 18 Jul 2025 16:59:27 GMT
            Vary:
              schema:
                type: string
                example: Origin
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: 30d548cb-caa5-4866-bdbe-22c7e93f745a
                  status:
                    type: string
                    example: success
              examples:
                sucess:
                  value:
                    data: 30d548cb-caa5-4866-bdbe-22c7e93f745a
                    status: success
        '400':
          description: error
          headers:
            Content-Length:
              schema:
                type: string
                example: '145'
            Date:
              schema:
                type: string
                example: Fri, 18 Jul 2025 17:00:26 GMT
            Vary:
              schema:
                type: string
                example: Origin
          content:
            application/json:
              schema:
                type: object
                properties:
                  details:
                    type: number
                    example: 400
                  error:
                    type: string
                    example: >-
                      Key: 'CreateSubscriptionRequest.Price' Error:Field
                      validation for 'Price' failed on the 'required' tag
                  status:
                    type: string
                    example: error
              examples:
                error:
                  value:
                    details: 400
                    error: >-
                      Key: 'CreateSubscriptionRequest.Price' Error:Field
                      validation for 'Price' failed on the 'required' tag
                    status: error
    delete:
      summary: delete subscriber
      description: delete subscriber
      operationId: deleteSubscriber
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
                    example: 688164ab-354a-4908-ae66-5a3d53908361
                  example:
                    - 688164ab-354a-4908-ae66-5a3d53908361
                    - 1ba5eb71-94b9-4f33-a2f8-77b85f52c9c9
            examples:
              delete subscriber:
                value:
                  ids:
                    - 688164ab-354a-4908-ae66-5a3d53908361
                    - 1ba5eb71-94b9-4f33-a2f8-77b85f52c9c9
      responses:
        '200':
          description: success
          headers:
            Content-Length:
              schema:
                type: string
                example: '63'
            Date:
              schema:
                type: string
                example: Fri, 18 Jul 2025 17:31:25 GMT
            Vary:
              schema:
                type: string
                example: Origin
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: All subscribers have been deleted
                  status:
                    type: string
                    example: success
              examples:
                success:
                  value:
                    data: All subscribers have been deleted
                    status: success
  /subscribe/cost:
    get:
      summary: agree subscribe
      description: agree subscribe
      operationId: agreeSubscribe
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
            example: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
        - name: service_name
          in: query
          schema:
            type: string
            example: Netflix Premium12
        - name: start_date
          in: query
          schema:
            type: string
            example: 01-2024
        - name: end_date
          in: query
          schema:
            type: string
            example: 02-2025
      responses:
        '200':
          description: success
          headers:
            Content-Length:
              schema:
                type: string
                example: '34'
            Date:
              schema:
                type: string
                example: Fri, 18 Jul 2025 19:44:29 GMT
            Vary:
              schema:
                type: string
                example: Origin
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: number
                    example: 100000
                  status:
                    type: string
                    example: success
              examples:
                success:
                  value:
                    data: 100000
                    status: success
        '400':
          description: error
          headers:
            Content-Length:
              schema:
                type: string
                example: '65'
            Date:
              schema:
                type: string
                example: Fri, 18 Jul 2025 19:44:52 GMT
            Vary:
              schema:
                type: string
                example: Origin
          content:
            application/json:
              schema:
                type: object
                properties:
                  details:
                    type: number
                    example: 400
                  error:
                    type: string
                    example: failed to get ent_date
                  status:
                    type: string
                    example: error
              examples:
                error:
                  value:
                    details: 400
                    error: failed to get ent_date
                    status: error
  /subscribes:
    get:
      summary: get subscribes
      description: get subscribes
      operationId: getSubscribes
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
        - name: service_name
          in: query
          schema:
            type: string
            example: Netflix Premium12
        - name: limit
          in: query
          schema:
            type: string
            example: '10'
        - name: offset
          in: query
          schema:
            type: string
            example: '0'
      responses:
        '200':
          description: success
          headers:
            Date:
              schema:
                type: string
                example: Fri, 18 Jul 2025 19:55:49 GMT
            Transfer-Encoding:
              schema:
                type: string
                example: chunked
            Vary:
              schema:
                type: string
                example: Origin
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        end_date:
                          type: string
                          example: '2024-12-01T00:00:00Z'
                        id:
                          type: string
                          example: af7a4ff5-6ce9-48d8-bf4e-ea28b005fadb
                        price:
                          type: number
                          example: 15
                        service_name:
                          type: string
                          example: Netflix Premium
                        start_date:
                          type: string
                          example: '2024-01-01T00:00:00Z'
                        user_id:
                          type: string
                          example: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                    example:
                      - end_date: '2024-12-01T00:00:00Z'
                        id: af7a4ff5-6ce9-48d8-bf4e-ea28b005fadb
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: fd57aa6f-88a7-4b04-9477-2bdeb3f3f20d
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: a470dc53-2658-428f-8ac7-0c91f1e77ae2
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: 6eed7d72-8a00-49dc-b5a0-1237959cfef7
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: 0577ae90-6df1-4d48-833b-a0cc821d2bef
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: cd904cbf-d629-4b2f-9b4c-b08570b22bbd
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: 411fe711-e2bd-4069-b070-f8078dc6f931
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: d5760810-5ae6-4eb4-8947-92eed1da3eaa
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '0001-01-01T00:00:00Z'
                        id: 31110cd2-0ed0-482f-9002-efafd17c003a
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '0001-01-01T00:00:00Z'
                        id: 37df5f72-8cea-4b81-9b82-b8997fc277f5
                        price: 15
                        service_name: Netflix Premium
                        start_date: '0001-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                  meta:
                    type: number
                    example: 200
                  status:
                    type: string
                    example: success
              examples:
                success:
                  value:
                    data:
                      - end_date: '2024-12-01T00:00:00Z'
                        id: af7a4ff5-6ce9-48d8-bf4e-ea28b005fadb
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: fd57aa6f-88a7-4b04-9477-2bdeb3f3f20d
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: a470dc53-2658-428f-8ac7-0c91f1e77ae2
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: 6eed7d72-8a00-49dc-b5a0-1237959cfef7
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: 0577ae90-6df1-4d48-833b-a0cc821d2bef
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: cd904cbf-d629-4b2f-9b4c-b08570b22bbd
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: 411fe711-e2bd-4069-b070-f8078dc6f931
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '2024-12-01T00:00:00Z'
                        id: d5760810-5ae6-4eb4-8947-92eed1da3eaa
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '0001-01-01T00:00:00Z'
                        id: 31110cd2-0ed0-482f-9002-efafd17c003a
                        price: 15
                        service_name: Netflix Premium
                        start_date: '2024-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                      - end_date: '0001-01-01T00:00:00Z'
                        id: 37df5f72-8cea-4b81-9b82-b8997fc277f5
                        price: 15
                        service_name: Netflix Premium
                        start_date: '0001-01-01T00:00:00Z'
                        user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                    meta: 200
                    status: success
        '400':
          description: error
          headers:
            Content-Length:
              schema:
                type: string
                example: '63'
            Date:
              schema:
                type: string
                example: Fri, 18 Jul 2025 19:57:38 GMT
            Vary:
              schema:
                type: string
                example: Origin
          content:
            application/json:
              schema:
                type: object
                properties:
                  details:
                    type: number
                    example: 400
                  error:
                    type: string
                    example: failed to get offset
                  status:
                    type: string
                    example: error
              examples:
                error:
                  value:
                    details: 400
                    error: failed to get offset
                    status: error
  /subscribe/df73e03e-ea64-415e-a1e0-222eeb137a1c:
    get:
      summary: get subscribe
      description: get subscribe
      operationId: getSubscribe
      responses:
        '200':
          description: success
          headers:
            Content-Length:
              schema:
                type: string
                example: '247'
            Date:
              schema:
                type: string
                example: Fri, 18 Jul 2025 19:58:37 GMT
            Vary:
              schema:
                type: string
                example: Origin
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      end_date:
                        type: string
                        example: '2024-12-01T00:00:00Z'
                      id:
                        type: string
                        example: df73e03e-ea64-415e-a1e0-222eeb137a1c
                      price:
                        type: number
                        example: 15
                      service_name:
                        type: string
                        example: Netflix Premium
                      start_date:
                        type: string
                        example: '2024-01-01T00:00:00Z'
                      user_id:
                        type: string
                        example: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                  meta:
                    type: number
                    example: 200
                  status:
                    type: string
                    example: success
              examples:
                success:
                  value:
                    data:
                      end_date: '2024-12-01T00:00:00Z'
                      id: df73e03e-ea64-415e-a1e0-222eeb137a1c
                      price: 15
                      service_name: Netflix Premium
                      start_date: '2024-01-01T00:00:00Z'
                      user_id: a1b2c3dx-e5f6-7890-g1h2-i3j4k5l6m7n2
                    meta: 200
                    status: success
        '400':
          description: error
          headers:
            Content-Length:
              schema:
                type: string
                example: '70'
            Date:
              schema:
                type: string
                example: Fri, 18 Jul 2025 19:59:16 GMT
            Vary:
              schema:
                type: string
                example: Origin
          content:
            application/json:
              schema:
                type: object
                properties:
                  details:
                    type: number
                    example: 400
                  error:
                    type: string
                    example: params id is not valid uuid
                  status:
                    type: string
                    example: error
              examples:
                error:
                  value:
                    details: 400
                    error: params id is not valid uuid
                    status: error
tags: []